writer-harness/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "writing-harness"
version = "0.1.0"
description = "CLI tool for enforcing continuity and style in serialized narratives"
readme = "README.md"
license = {text = "MIT"}
authors = [{name = "Your Name", email = "you@example.com"}]
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "rich>=13.0",
    "python-dotenv>=1.0",
    "PyYAML>=6.0",
    "anthropic>=0.28",
    "openai>=1.0",
    "rapidfuzz>=3.0",
    "click>=8.0",
]

[project.scripts]
harness = "harness.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1",
]
